Revert "Multi Team Spawns"
This reverts commit 8b1885d07686409592f231184a30f049edb08fa8.
