+++ locale/cz.yml
@@ -52,9 +52,7 @@ success:
     saved: "Hra byla uspesne ulozena!"
     bedset: "You sucessfully set the respawn block of team $team$!"
     regionset: "Oblast $location$ pro hru $game$ uspesne nastavena!"
-    spawnadd: "Spawn location for Team $team$ was added successfully!"
-    spawnremove: "Spawn location for Team $team$ was removed successfully!"
-    spawnclear: " All Spawn locations for Team $team$ were cleared successfully!"
+    spawnset: "Spawn lokace pro tym $team$ uspesne nastavena!"
     spawnerset: "Spawn surovin $name$ uspesne nastaven!"
     stopped: "Hra byla zastavena!"
     lobbyset: "Lobby pro hru bylo nastaveno!"
@@ -113,15 +111,15 @@ ingame:
         beddestroyed: "$player$ destroyed bed of team $team$!"
     specials:
         rescue-platform:
-            left: Musis pockat &c$time$&f sekund(u) pred pouzitim dalsi zachrane platformy!
+            left: "&cMusis pockat &e$time$&f sekund(u) pred pouzitim dalsi zachrane platformy!"
         trap:
             trapped: "&eNekdo stoupl do &cpasti&e tveho tymu!"
         protection-wall:
             left: "Musis pockat &c$time$&f sekund pred pouzitim!"
-            not-usable-here: "V tomto miste nemuzes zed pouzit!"
+            not-usable-here: "Tento predmet zde nemuzes pouzit!"
         warp-powder:
             cancelled: "&cTeleportace zrusena!"
-            start: "Budes teleportovan za &c$time$&f sekund! &4NEHYBEJ SE!"
+            start: "Budes teleportovan za &c$time$&f sekund!"
             cancel: "&4Zrusit Teleport"
 lobby:
     countdown: "Hra zacne za $sec$ sekund!"
@@ -129,8 +127,8 @@ lobby:
     cancelcountdown:
         TEAMS_HAVE_PLAYERS: "Minimalne jeden tym nema hrace! Odpocet byl zastaven.!"
         PLAYERS_IN_GAME: "Je potreba vice hracu! Odpocet byl zastaven."
-        ENOUGH_TEAMS_AND_PLAYERS: "Pouze jeden tym ma hrace nebo je potreba vice hracu! Odpovet byl zastaven."
-    chooseteam: "Vyber si team"
+        ENOUGH_TEAMS_AND_PLAYERS: "Pouze jeden tym ma hrace nebo je potreba vice hracu! Odpocet byl zastaven."
+    chooseteam: "Vyber si tym"
     startgame: "Spustit hru"
     gamefull: "Hra je plna!"
     teamjoined: "Uspesne jsi se pripojil do tymu $team$"
@@ -188,15 +186,9 @@ commands:
     setregion:
         name: "Nastavit oblast"
         desc: "Nastavi hranice hry"
-    addspawn:
-        name: "Adds a team spawn"
-        desc: "Adds a spawn of the given team"
-    removespawn:
-        name: "Removes a team spawn"
-        desc: "Removes a spawn of the given team"
-    clearspawns:
-        name: "Clears all team spawns"
-        desc: "Clears all spawn of the given team"
+    setspawn:
+        name: "Nastavit spawn tymu"
+        desc: "Nastavi misto pro spawn urciteho tymu"
     setspawner:
         name: "Nastavit spawner"
         desc: "Nastavi misto pro spawnovani surovin"
